backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
numpy==1.26.3
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
